fix(media): move mediaConn to base socket + pre-fetch on success + thread host through app-state/history downloads
Fixes the issue where app-state-sync external-blob downloads and
history-sync downloads hit \`mmg.whatsapp.net\` even when WhatsApp
routed the socket to a different CDN shard, producing 400s under
\`extractSyncdPatches\` / \`downloadAndProcessHistorySyncNotification\`.

Three problems stacked:

1. The \`mediaConn\` / \`mediaHost\` / \`refreshMediaConn\` machinery
   lived in \`messages-send.ts\`, which is ABOVE \`chats.ts\` in the
   socket layering. \`chats.ts\` could not access \`getMediaHost\` from
   inside its own \`resyncAppState\` / \`appPatch\` code paths.

2. The internal \`extractSyncdPatches\` → \`downloadExternalBlob\` →
   \`downloadContentFromMessage\` chain in \`chat-utils.ts\` (and the
   parallel \`downloadHistory\` chain in \`history.ts\`) didn't accept a
   host parameter at all, so they always fell back to
   \`DEF_MEDIA_HOST\` via \`getUrlFromDirectPath\`.

3. \`refreshMediaConn\` was only fired on the first upload. App-state
   sync and history-sync notifications arrive BEFORE any upload, so
   \`mediaHost\` stayed at the \`DEF_MEDIA_HOST\` default forever for
   pure-receive workloads (read-only bots, history-only consumers,
   benchmarks).

Fixes:

- Moved \`mediaConn\` / \`mediaHost\` / \`refreshMediaConn\` /
  \`safeAwaitMediaConn\` / \`mediaConnMutex\` / \`getMediaHost\` to
  \`socket.ts\` (base socket). Now exposed on every layer above.
  \`messages-send.ts\` destructures them from \`sock\` rather than
  defining its own copies. The previous duplicate was removed; the
  registerSocketEndHandler clearing of \`mediaConn\` also moved up.

- Added an optional \`host?: string\` parameter to:
  - \`chat-utils.ts\`: \`extractSyncdPatches\`, \`downloadExternalBlob\`,
    \`downloadExternalPatch\`, \`decodePatches\`
  - \`history.ts\`: \`downloadHistory\`,
    \`downloadAndProcessHistorySyncNotification\`
  These thread the host into the existing \`downloadContentFromMessage\`
  call which already accepted \`opts.host\`. Back-compat: callers that
  don't pass \`host\` get the previous \`DEF_MEDIA_HOST\` fallback.

- Added \`getMediaHost?: () => string\` to \`ProcessMessageContext\` so
  history-sync downloads inside \`processMessage\` can use the
  per-socket host.

- Call sites in \`chats.ts\` (extractSyncdPatches at line ~675,
  decodePatches at lines ~698 and ~1003) and \`process-message.ts\`
  (downloadAndProcessHistorySyncNotification at line ~478) now pass
  \`getMediaHost()\` (or the getter via context).

- \`ws.on('CB:success', …)\` in \`socket.ts\` now eagerly calls
  \`await refreshMediaConn()\` once after the login sequence
  completes, so \`mediaHost\` is populated BEFORE the first
  app-state-sync or history-sync arrives. Failure is non-fatal —
  logged as a warn; downloads fall back to \`DEF_MEDIA_HOST\` and
  upload paths still re-fetch on demand.

Verification: 50 suites pass (486 + 5 todo), 0 lint errors, types clean.

The bench log that triggered this (downloads hitting
mmg.whatsapp.net and 400-ing under app-state sync) should clear once
the lib is rebuilt against this commit.
