+++ src/Socket/messages-send.ts
@@ -4,7 +4,6 @@ import { proto } from '../../WAProto/index.js'
 import { DEFAULT_CACHE_TTLS, WA_DEFAULT_EPHEMERAL } from '../Defaults'
 import type {
 	AnyMessageContent,
-	MediaConnInfo,
 	MessageReceiptType,
 	MessageRelayOptions,
 	MiscMessageGenerationOptions,
@@ -18,7 +17,6 @@ import {
 	assertMeId,
 	bindWaitForEvent,
 	decryptMediaRetryData,
-	DEF_MEDIA_HOST,
 	downloadContentFromMessage,
 	downloadMediaMessage,
 	encodeNewsletterMessage,
@@ -54,8 +52,6 @@ import {
 	type BinaryNode,
 	type BinaryNodeAttributes,
 	type FullJid,
-	getBinaryNodeChild,
-	getBinaryNodeChildren,
 	isHostedLidUser,
 	isHostedPnUser,
 	isJidBot,
@@ -96,7 +92,9 @@ export const makeMessagesSocket = (config: SocketConfig) => {
 		sendNode,
 		groupMetadata,
 		groupToggleEphemeral,
-		registerSocketEndHandler
+		registerSocketEndHandler,
+		refreshMediaConn,
+		getMediaHost
 	} = sock
 
 	const getLIDForPN = signalRepository.lidMapping.getLIDForPN.bind(signalRepository.lidMapping)
@@ -123,79 +121,11 @@ export const makeMessagesSocket = (config: SocketConfig) => {
 	// Prevent race conditions in Signal session encryption by user
 	const encryptionMutex = makeKeyedMutex()
 
-	let mediaConn: Promise<MediaConnInfo> | undefined
-	/** Per-socket media host; updated whenever media_conn is fetched. Defaults to the public WhatsApp host. */
-	let mediaHost: string = DEF_MEDIA_HOST
-	/**
-	 * Single-flight guard for the media-conn fetch (Stage 9). Two concurrent
-	 * callers used to both observe `mediaConn` as expired and both reassign
-	 * the in-flight promise, leaking the first fetch's result. The mutex
-	 * ensures only one refresh runs at a time; subsequent callers await the
-	 * one in-flight or reuse the freshly-cached value.
-	 */
-	const mediaConnMutex = makeMutex()
-	/**
-	 * Safely await the cached `mediaConn` promise. If a previous fetch
-	 * stored a REJECTED promise, awaiting it again would just rethrow
-	 * forever and poison every subsequent caller with the stale failure
-	 * — there'd be no way to retry. Clear the cache slot on rejection so
-	 * the next caller falls through to a fresh `media_conn` query.
-	 */
-	const safeAwaitMediaConn = async (): Promise<MediaConnInfo | undefined> => {
-		if (!mediaConn) return undefined
-		try {
-			return await mediaConn
-		} catch (err) {
-			logger.warn?.({ err }, 'previous media_conn fetch failed, will retry')
-			mediaConn = undefined
-			return undefined
-		}
-	}
-
-	const refreshMediaConn = async (forceGet = false): Promise<MediaConnInfo> => {
-		const cached = await safeAwaitMediaConn()
-		if (cached && !forceGet && new Date().getTime() - cached.fetchDate.getTime() <= cached.ttl * 1000) {
-			return cached
-		}
-
-		return mediaConnMutex.mutex(async () => {
-			// Re-check inside the lock: another caller may have refreshed.
-			const after = await safeAwaitMediaConn()
-			if (after && !forceGet && new Date().getTime() - after.fetchDate.getTime() <= after.ttl * 1000) {
-				return after
-			}
-
-			mediaConn = (async () => {
-				const result = await query({
-					tag: 'iq',
-					attrs: {
-						type: 'set',
-						xmlns: 'w:m',
-						to: S_WHATSAPP_NET
-					},
-					content: [{ tag: 'media_conn', attrs: {} }]
-				})
-				const mediaConnNode = getBinaryNodeChild(result, 'media_conn')!
-				const node: MediaConnInfo = {
-					hosts: getBinaryNodeChildren(mediaConnNode, 'host').map(({ attrs }) => ({
-						hostname: attrs.hostname!,
-						maxContentLengthBytes: +attrs.maxContentLengthBytes!
-					})),
-					auth: mediaConnNode.attrs.auth!,
-					ttl: +mediaConnNode.attrs.ttl!,
-					fetchDate: new Date()
-				}
-				logger.debug('fetched media conn')
-				if (node.hosts[0]) {
-					mediaHost = node.hosts[0].hostname
-				}
-
-				return node
-			})()
-
-			return mediaConn
-		})
-	}
+	// `mediaConn` / `mediaHost` / `refreshMediaConn` moved to the base
+	// socket in Stage 11 so chats.ts (which is BELOW this layer in the
+	// socket layering) can also route its app-state / history-sync
+	// downloads through the correct CDN shard. We just pull the helpers
+	// out of `sock`.
 
 	/**
 	 * generic send receipt function
@@ -1324,7 +1254,9 @@ export const makeMessagesSocket = (config: SocketConfig) => {
 			userDevicesCache.close()
 		}
 
-		mediaConn = undefined
+		// `mediaConn` clearing now lives in the base socket's own
+		// teardown — moved alongside the rest of the media-conn
+		// machinery in Stage 11.
 		if (messageRetryManager) {
 			messageRetryManager.clear()
 		}
@@ -1346,10 +1278,10 @@ export const makeMessagesSocket = (config: SocketConfig) => {
 	 */
 	const downloadMediaMessageWithHost = <Type extends 'buffer' | 'stream'>(
 		...[message, type, options, ctx]: Parameters<typeof downloadMediaMessage<Type>>
-	) => downloadMediaMessage<Type>(message, type, { host: mediaHost, ...options }, ctx)
+	) => downloadMediaMessage<Type>(message, type, { host: getMediaHost(), ...options }, ctx)
 
 	const downloadContentFromMessageWithHost: typeof downloadContentFromMessage = (message, type, opts = {}) =>
-		downloadContentFromMessage(message, type, { host: mediaHost, ...opts })
+		downloadContentFromMessage(message, type, { host: getMediaHost(), ...opts })
 
 	return {
 		...sock,
@@ -1362,10 +1294,12 @@ export const makeMessagesSocket = (config: SocketConfig) => {
 		sendReceipts,
 		readMessages,
 		refreshMediaConn,
-		// Function (not getter) so the spread in chats.ts preserves the live closure binding.
-		getMediaHost: () => mediaHost,
+		// `getMediaHost` is already re-exported from the base socket via
+		// the `...sock` spread above. Keeping the explicit re-export here
+		// so the typed return shape from this layer is unchanged.
+		getMediaHost,
 		/** Build a media URL using the socket's current `media_conn` host. */
-		getMediaUrl: (directPath: string) => getUrlFromDirectPath(directPath, mediaHost),
+		getMediaUrl: (directPath: string) => getUrlFromDirectPath(directPath, getMediaHost()),
 		/**
 		 * Download a media message via the per-socket `media_conn` host. Same
 		 * signature as the standalone utility; pass `options.host` to
@@ -1411,7 +1345,7 @@ export const makeMessagesSocket = (config: SocketConfig) => {
 								}
 
 								content.directPath = media.directPath
-								content.url = getUrlFromDirectPath(content.directPath!, mediaHost)
+								content.url = getUrlFromDirectPath(content.directPath!, getMediaHost())
 
 								logger.debug({ directPath: media.directPath, key: result.key }, 'media update successful')
 							} catch (err: any) {
