+++ src/Utils/chat-utils.ts
@@ -374,7 +374,20 @@ export const decodeSyncdPatch = async (
 	return result
 }
 
-export const extractSyncdPatches = async (result: BinaryNode, options: RequestInit) => {
+/**
+ * Decode the `<sync>` IQ result returned by the app-state-sync endpoint.
+ * When a collection's mutations are too large to fit inline, they're
+ * carried as an `ExternalBlobReference`; we fetch the blob via
+ * `downloadExternalBlob` — which in turn calls
+ * `downloadContentFromMessage` to hit the WhatsApp media CDN.
+ *
+ * `host` is the optional per-socket media-CDN hostname (from
+ * `sock.getMediaHost()`). Passing it threads through to the download so
+ * we hit the correct shard for this socket's region; omit it and the
+ * download falls back to `DEF_MEDIA_HOST` (`mmg.whatsapp.net`) which
+ * may 400 on regions WhatsApp routed to a different host.
+ */
+export const extractSyncdPatches = async (result: BinaryNode, options: RequestInit, host?: string) => {
 	const syncNode = getBinaryNodeChild(result, 'sync')
 	const collectionNodes = getBinaryNodeChildren(syncNode, 'collection')
 
@@ -400,7 +413,7 @@ export const extractSyncdPatches = async (result: BinaryNode, options: RequestIn
 				}
 
 				const blobRef = proto.ExternalBlobReference.decode(snapshotNode.content as Buffer)
-				const data = await downloadExternalBlob(blobRef, options)
+				const data = await downloadExternalBlob(blobRef, options, host)
 				snapshot = proto.SyncdSnapshot.decode(data)
 			}
 
@@ -426,8 +439,8 @@ export const extractSyncdPatches = async (result: BinaryNode, options: RequestIn
 	return final
 }
 
-export const downloadExternalBlob = async (blob: proto.IExternalBlobReference, options: RequestInit) => {
-	const stream = await downloadContentFromMessage(blob, 'md-app-state', { options })
+export const downloadExternalBlob = async (blob: proto.IExternalBlobReference, options: RequestInit, host?: string) => {
+	const stream = await downloadContentFromMessage(blob, 'md-app-state', { options, host })
 	const bufferArray: Buffer[] = []
 	for await (const chunk of stream) {
 		bufferArray.push(chunk)
@@ -436,8 +449,12 @@ export const downloadExternalBlob = async (blob: proto.IExternalBlobReference, o
 	return Buffer.concat(bufferArray)
 }
 
-export const downloadExternalPatch = async (blob: proto.IExternalBlobReference, options: RequestInit) => {
-	const buffer = await downloadExternalBlob(blob, options)
+export const downloadExternalPatch = async (
+	blob: proto.IExternalBlobReference,
+	options: RequestInit,
+	host?: string
+) => {
+	const buffer = await downloadExternalBlob(blob, options, host)
 	const syncData = proto.SyncdMutations.decode(buffer)
 	return syncData
 }
@@ -507,7 +524,8 @@ export const decodePatches = async (
 	options: RequestInit,
 	minimumVersionNumber?: number,
 	logger?: ILogger,
-	validateMacs = true
+	validateMacs = true,
+	host?: string
 ) => {
 	const newState: LTHashState = {
 		...initial,
@@ -520,7 +538,7 @@ export const decodePatches = async (
 		const { version, keyId, snapshotMac } = syncd
 		if (syncd.externalMutations) {
 			logger?.trace({ name, version }, 'downloading external patch')
-			const ref = await downloadExternalPatch(syncd.externalMutations, options)
+			const ref = await downloadExternalPatch(syncd.externalMutations, options, host)
 			logger?.debug({ name, version, mutations: ref.mutations.length }, 'downloaded external patch')
 			syncd.mutations?.push(...ref.mutations)
 		}
